requirements.txt
mysql-connector-python==9.1.0
streamlit==1.39.0
pandas==2.2.3
numpy==2.1.2